Worked a bit on coding standards

+++ tests/testDetection.cc
@@ -6,7 +6,7 @@
 using namespace std;
 using namespace lsst::fw;
 
-typedef uint16 MaskT;
+typedef uint8 MaskT;
 typedef float ImageT;
 typedef double KernelT;
 typedef double FuncT;
